+++ Examples/Map+Positioning/Sources/ViewControllers/InitialViewController.swift
@@ -9,6 +9,7 @@
 import RxCocoa
 import RxSwift
 import UIKit
+import WemapCoreSDK
 import WemapMapSDK
 import WemapPositioningSDKPolestar
 import WemapPositioningSDKVPSARKit